requirements.txt
instaloader==4.10
requests==2.31.0
python-telegram-bot==20.7
